Move note driver from drivers/syslog to drivers/note
it's better to put the note transport layer into a common folder
