+++ arch/z80/src/common/z80_initialize.c
@@ -34,7 +34,7 @@
 #include <nuttx/net/loopback.h>
 #include <nuttx/net/telnet.h>
 #include <nuttx/net/tun.h>
-#include <nuttx/syslog/note_driver.h>
+#include <nuttx/note/note_driver.h>
 #include <nuttx/syslog/syslog_console.h>
 #include <nuttx/serial/pty.h>
 #include <nuttx/crypto/crypto.h>
