+++ drivers/Kconfig
@@ -96,6 +96,7 @@ source drivers/modem/Kconfig
 source drivers/mtd/Kconfig
 source drivers/eeprom/Kconfig
 source drivers/net/Kconfig
+source drivers/note/Kconfig
 source drivers/pipes/Kconfig
 source drivers/power/Kconfig
 source drivers/rptun/Kconfig
