+++ drivers/Makefile
@@ -56,6 +56,7 @@ include modem/Make.defs
 include mtd/Make.defs
 include eeprom/Make.defs
 include net/Make.defs
+include note/Make.defs
 include pipes/Make.defs
 include power/Make.defs
 include rptun/Make.defs
