+++ drivers/note/Kconfig
@@ -1,0 +1,17 @@
+#
+# For a description of the syntax of this configuration file,
+# see the file kconfig-language.txt in the NuttX tools repository.
+#
+
+menu "Note Driver Support"
+
+config DRIVER_NOTE
+	bool "Scheduler instrumentation driver"
+	default n
+	depends on SCHED_INSTRUMENTATION_BUFFER && SCHED_NOTE_GET
+	---help---
+		Enable building a serial driver that can be used by an application
+		to read data from the in-memory, scheduler instrumentation "note"
+		buffer.
+
+endmenu
