+++ drivers/syslog/Kconfig
@@ -68,15 +68,6 @@ config RAMLOG_OVERWRITE
 
 endif
 
-config DRIVER_NOTE
-	bool "Scheduler instrumentation driver"
-	default n
-	depends on SCHED_INSTRUMENTATION_BUFFER && SCHED_NOTE_GET
-	---help---
-		Enable building a serial driver that can be used by an application
-		to read data from the in-memory, scheduler instrumentation "note"
-		buffer.
-
 config SYSLOG_BUFFER
 	bool "Use buffered output"
 	default n
